Dialogue changes
Dialogue now is dynamic but it changes too quickly for you to read
anything. I've disabled it disappearing for now, but it now also
reappears and disappears (also too quickly for you to see). Will need to
add a coroutine to fix that, but with that fix, the entire dialogue
system will be working I think? Besides changing portraits.