Save Function now partly broken
-Quit menu now added upon pressing quit button from resume button
-Updated Load menu text to say something like "choose a file"
-Save function works...sorta. Save function works if you start the game,
go in, save, head back into the editor, load the game. It doesn't work
where you go into the game, quit using the ingame menu and then try to
load. It has to do with the fact that the GameManager reference changes
how it works by becoming a DontDestroyOnLoad object or whatever. Will
fix tomorrow.